pyproject.toml
[tool.poetry]
name = "avangard-test"
version = "0.1.0"
description = ""
authors = ["RG1ee <niwows@yandex.ru>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
flask-restful = "^0.3.10"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.5"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
flask-jwt-extended = "^4.5.3"
flask-marshmallow = "^0.15.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"
ruff = "^0.1.0"
black = "^23.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
ignore = ["F403", "F405", "F811"]
exclude = [
    "migrations",
    ".direnv",
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".tox",
    ".venv",
    "venv",
    "__pypackages__",
    "dist",
    "node_modules",
]
line-length = 120

[tool.poetry.scripts]
runserver = "src.app:main"
